Change link
The Prindle Institute now has their own website, so old links to them
are now broken. This changes the link to one via the Wayback machine for
now, but it might be worth keeping an eye on their new
website (https://www.prindleinstitute.org/) in case they decide to port
old things over.