Don't display psych contributions
The news page that I linked to on CRAN seems to have disappeared for the
psych package. I'm not sure if this is temporary or if the list of news
items have been removed from the package, but this means that the page
which listed my contributions is no longer available. Since the
contributions weren't that great anyway, I've decided to remove it for
now, but I should keep an eye on this in case it should be brought back.